again

+++ philo_one copy/srcs/utils.c
@@ -6,13 +6,13 @@
 /*   By: fredrika <fredrika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/10 22:25:54 by fredrika          #+#    #+#             */
-/*   Updated: 2020/03/27 12:27:41 by fredrikalindh    ###   ########.fr       */
+/*   Updated: 2020/03/27 12:25:49 by fredrikalindh    ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <philo.h>
 
-u_int64_t	get_time(void)
+unsigned long long	get_time(void)
 {
 	static struct timeval	start = {0, 0};
 	struct timeval			tv;
@@ -20,6 +20,6 @@ u_int64_t	get_time(void)
 	if (!start.tv_sec)
 		gettimeofday(&start, NULL);
 	gettimeofday(&tv, NULL);
-	return (((tv.tv_sec - start.tv_sec) * (uint64_t)1000) +
-	((tv.tv_usec - start.tv_usec) / 1000));
+	return (((tv.tv_sec - start.tv_sec) * (unsigned long long)1000) +
+			((tv.tv_usec - start.tv_usec) / 1000));
 }
